+++ include/nanofmt/CMakeLists.txt
@@ -1,8 +1,8 @@
 target_sources(nanofmt PRIVATE
     "buffer.h"
+    "charconv.h"
     "config.h"
     "format.h"
     "format.inl"
     "std_string.h"
-    "to_chars.h"
 )
