+++ source/CMakeLists.txt
@@ -1,9 +1,9 @@
 include(fetch_dragonbox)
 
 target_sources(nanofmt PRIVATE
+    "charconv.cpp"
     "format.cpp"
     "numeric_utils.h"
     "parse_utils.h"
-    "to_chars.cpp"
 )
 target_include_directories(nanofmt SYSTEM PRIVATE ${nanofmt_dragonbox_SOURCE_DIR}/include)
