+++ tests/CMakeLists.txt
@@ -2,11 +2,12 @@ include(fetch_catch2)
 
 add_executable(nanofmt_test)
 target_sources(nanofmt_test PRIVATE
+    "test_charconv.cpp"
     "test_format.cpp"
     "test_format_args.cpp"
-    "test_to_chars.cpp"
     "test_utils.h"
- "test_format_value.cpp")
+    "test_format_value.cpp"
+)
 target_link_libraries(nanofmt_test PRIVATE
     nanofmt
     Catch2::Catch2WithMain
