+++ tests/test_utils.h
@@ -1,7 +1,7 @@
 // Copyright (c) Sean Middleditch and contributors. See accompanying LICENSE.md for copyright details.
 
+#include "nanofmt/charconv.h"
 #include "nanofmt/format.h"
-#include "nanofmt/to_chars.h"
 
 #include <cstring>
 #include <iostream>
